packages/backend/requirements.txt
web3==6.20.0
eth-account==0.11.0
python-dotenv==1.0.0
aiohttp==3.9.0
goat-sdk==0.1.6
langchain==0.3.0
langchain-openai==0.2.0
eth-abi==5.1.0
eth-utils==5.1.0